Split Kokkos_Parallel_Team.hpp
