+++ core/src/SYCL/Kokkos_SYCL_ParallelReduce_Team.hpp
@@ -14,518 +14,21 @@
 //
 //@HEADER
 
-#ifndef KOKKOS_SYCL_PARALLEL_TEAM_HPP
-#define KOKKOS_SYCL_PARALLEL_TEAM_HPP
+#ifndef KOKKOS_SYCL_PARALLEL_REDUCE_TEAM_HPP
+#define KOKKOS_SYCL_PARALLEL_REDUCE_TEAM_HPP
 
 #include <Kokkos_Parallel.hpp>
 
-#include <SYCL/Kokkos_SYCL_WorkgroupReduction.hpp>
 #include <SYCL/Kokkos_SYCL_Team.hpp>
+#include <SYCL/Kokkos_SYCL_TeamPolicy.hpp>
+#include <SYCL/Kokkos_SYCL_WorkgroupReduction.hpp>
 
 #include <vector>
 
-namespace Kokkos {
-namespace Impl {
-template <typename... Properties>
-class TeamPolicyInternal<Kokkos::Experimental::SYCL, Properties...>
-    : public PolicyTraits<Properties...> {
- public:
-  using execution_policy = TeamPolicyInternal;
-
-  using traits = PolicyTraits<Properties...>;
-
-  template <typename ExecSpace, typename... OtherProperties>
-  friend class TeamPolicyInternal;
-
- private:
-  typename traits::execution_space m_space;
-  int m_league_size;
-  int m_team_size;
-  int m_vector_length;
-  size_t m_team_scratch_size[2];
-  size_t m_thread_scratch_size[2];
-  int m_chunk_size;
-  bool m_tune_team_size;
-  bool m_tune_vector_length;
-
- public:
-  using execution_space = Kokkos::Experimental::SYCL;
-
-  template <class... OtherProperties>
-  TeamPolicyInternal(TeamPolicyInternal<OtherProperties...> const& p) {
-    m_league_size            = p.m_league_size;
-    m_team_size              = p.m_team_size;
-    m_vector_length          = p.m_vector_length;
-    m_team_scratch_size[0]   = p.m_team_scratch_size[0];
-    m_team_scratch_size[1]   = p.m_team_scratch_size[1];
-    m_thread_scratch_size[0] = p.m_thread_scratch_size[0];
-    m_thread_scratch_size[1] = p.m_thread_scratch_size[1];
-    m_chunk_size             = p.m_chunk_size;
-    m_space                  = p.m_space;
-    m_tune_team_size         = p.m_tune_team_size;
-    m_tune_vector_length     = p.m_tune_vector_length;
-  }
-
-  template <typename FunctorType>
-  int team_size_max(FunctorType const& f, ParallelForTag const&) const {
-    return internal_team_size_max_for(f);
-  }
-
-  template <class FunctorType>
-  inline int team_size_max(const FunctorType& f,
-                           const ParallelReduceTag&) const {
-    return internal_team_size_max_reduce<void>(f);
-  }
-
-  template <class FunctorType, class ReducerType>
-  inline int team_size_max(const FunctorType& f, const ReducerType& /*r*/,
-                           const ParallelReduceTag&) const {
-    return internal_team_size_max_reduce<typename ReducerType::value_type>(f);
-  }
-
-  template <typename FunctorType>
-  int team_size_recommended(FunctorType const& f, ParallelForTag const&) const {
-    return internal_team_size_recommended_for(f);
-  }
-
-  template <typename FunctorType>
-  inline int team_size_recommended(FunctorType const& f,
-                                   ParallelReduceTag const&) const {
-    return internal_team_size_recommended_reduce<void>(f);
-  }
-
-  template <class FunctorType, class ReducerType>
-  int team_size_recommended(FunctorType const& f, ReducerType const&,
-                            ParallelReduceTag const&) const {
-    return internal_team_size_recommended_reduce<
-        typename ReducerType::value_type>(f);
-  }
-  inline bool impl_auto_vector_length() const { return m_tune_vector_length; }
-  inline bool impl_auto_team_size() const { return m_tune_team_size; }
-  // FIXME_SYCL This is correct in most cases, but not necessarily in case a
-  // custom sycl::queue is used to initialize the execution space.
-  static int vector_length_max() {
-    std::vector<size_t> sub_group_sizes =
-        execution_space{}
-            .impl_internal_space_instance()
-            ->m_queue->get_device()
-            .template get_info<sycl::info::device::sub_group_sizes>();
-    return *std::max_element(sub_group_sizes.begin(), sub_group_sizes.end());
-  }
-
- private:
-  static int verify_requested_vector_length(int requested_vector_length) {
-    int test_vector_length =
-        std::min(requested_vector_length, vector_length_max());
-
-    // Allow only power-of-two vector_length
-    if (!(is_integral_power_of_two(test_vector_length))) {
-      int test_pow2 = 1;
-      while (test_pow2 < test_vector_length) test_pow2 <<= 1;
-      test_vector_length = test_pow2 >> 1;
-    }
-
-    return test_vector_length;
-  }
-
- public:
-  static int scratch_size_max(int level) {
-    return level == 0 ? 1024 * 32
-                      :           // FIXME_SYCL arbitrarily setting this to 32kB
-               20 * 1024 * 1024;  // FIXME_SYCL arbitrarily setting this to 20MB
-  }
-  inline void impl_set_vector_length(size_t size) { m_vector_length = size; }
-  inline void impl_set_team_size(size_t size) { m_team_size = size; }
-  int impl_vector_length() const { return m_vector_length; }
-
-  int team_size() const { return m_team_size; }
-
-  int league_size() const { return m_league_size; }
-
-  size_t scratch_size(int level, int team_size_ = -1) const {
-    if (team_size_ < 0) team_size_ = m_team_size;
-    return m_team_scratch_size[level] +
-           team_size_ * m_thread_scratch_size[level];
-  }
-
-  size_t team_scratch_size(int level) const {
-    return m_team_scratch_size[level];
-  }
-
-  size_t thread_scratch_size(int level) const {
-    return m_thread_scratch_size[level];
-  }
-
-  typename traits::execution_space space() const { return m_space; }
-
-  TeamPolicyInternal()
-      : m_space(typename traits::execution_space()),
-        m_league_size(0),
-        m_team_size(-1),
-        m_vector_length(0),
-        m_team_scratch_size{0, 0},
-        m_thread_scratch_size{0, 0},
-        m_chunk_size(vector_length_max()),
-        m_tune_team_size(false),
-        m_tune_vector_length(false) {}
-
-  /** \brief  Specify league size, request team size */
-  TeamPolicyInternal(const execution_space space_, int league_size_,
-                     int team_size_request, int vector_length_request = 1)
-      : m_space(space_),
-        m_league_size(league_size_),
-        m_team_size(team_size_request),
-        m_vector_length(
-            (vector_length_request > 0)
-                ? verify_requested_vector_length(vector_length_request)
-                : (verify_requested_vector_length(1))),
-        m_team_scratch_size{0, 0},
-        m_thread_scratch_size{0, 0},
-        m_chunk_size(vector_length_max()),
-        m_tune_team_size(bool(team_size_request <= 0)),
-        m_tune_vector_length(bool(vector_length_request <= 0)) {
-    // FIXME_SYCL Check that league size is permissible,
-    // https://github.com/intel/llvm/pull/4064
-
-    // Make sure total block size is permissible
-    if (m_team_size * m_vector_length >
-        static_cast<int>(
-            m_space.impl_internal_space_instance()->m_maxWorkgroupSize)) {
-      Impl::throw_runtime_exception(
-          std::string("Kokkos::TeamPolicy<SYCL> the team size is too large. "
-                      "Team size x vector length is " +
-                      std::to_string(m_team_size * m_vector_length) +
-                      " but must be smaller than ") +
-          std::to_string(
-              m_space.impl_internal_space_instance()->m_maxWorkgroupSize));
-    }
-  }
-
-  /** \brief  Specify league size, request team size */
-  TeamPolicyInternal(const execution_space space_, int league_size_,
-                     const Kokkos::AUTO_t& /* team_size_request */,
-                     int vector_length_request = 1)
-      : TeamPolicyInternal(space_, league_size_, -1, vector_length_request) {}
-  // FLAG
-  /** \brief  Specify league size and team size, request vector length*/
-  TeamPolicyInternal(const execution_space space_, int league_size_,
-                     int team_size_request,
-                     const Kokkos::AUTO_t& /* vector_length_request */
-                     )
-      : TeamPolicyInternal(space_, league_size_, team_size_request, -1)
-
-  {}
-
-  /** \brief  Specify league size, request team size and vector length*/
-  TeamPolicyInternal(const execution_space space_, int league_size_,
-                     const Kokkos::AUTO_t& /* team_size_request */,
-                     const Kokkos::AUTO_t& /* vector_length_request */
-
-                     )
-      : TeamPolicyInternal(space_, league_size_, -1, -1)
-
-  {}
-
-  TeamPolicyInternal(int league_size_, int team_size_request,
-                     int vector_length_request = 1)
-      : TeamPolicyInternal(typename traits::execution_space(), league_size_,
-                           team_size_request, vector_length_request) {}
-
-  TeamPolicyInternal(int league_size_,
-                     const Kokkos::AUTO_t& /* team_size_request */,
-                     int vector_length_request = 1)
-      : TeamPolicyInternal(typename traits::execution_space(), league_size_, -1,
-                           vector_length_request) {}
-
-  /** \brief  Specify league size and team size, request vector length*/
-  TeamPolicyInternal(int league_size_, int team_size_request,
-                     const Kokkos::AUTO_t& /* vector_length_request */
-
-                     )
-      : TeamPolicyInternal(typename traits::execution_space(), league_size_,
-                           team_size_request, -1)
-
-  {}
-
-  /** \brief  Specify league size, request team size and vector length*/
-  TeamPolicyInternal(int league_size_,
-                     const Kokkos::AUTO_t& /* team_size_request */,
-                     const Kokkos::AUTO_t& /* vector_length_request */
-
-                     )
-      : TeamPolicyInternal(typename traits::execution_space(), league_size_, -1,
-                           -1) {}
-
-  int chunk_size() const { return m_chunk_size; }
-
-  TeamPolicyInternal& set_chunk_size(typename traits::index_type chunk_size_) {
-    m_chunk_size = chunk_size_;
-    return *this;
-  }
-
-  /** \brief set per team scratch size for a specific level of the scratch
-   * hierarchy */
-  TeamPolicyInternal& set_scratch_size(int level,
-                                       PerTeamValue const& per_team) {
-    m_team_scratch_size[level] = per_team.value;
-    return *this;
-  }
-
-  /** \brief set per thread scratch size for a specific level of the scratch
-   * hierarchy */
-  TeamPolicyInternal& set_scratch_size(int level,
-                                       PerThreadValue const& per_thread) {
-    m_thread_scratch_size[level] = per_thread.value;
-    return *this;
-  }
-
-  /** \brief set per thread and per team scratch size for a specific level of
-   * the scratch hierarchy */
-  TeamPolicyInternal& set_scratch_size(int level, PerTeamValue const& per_team,
-                                       PerThreadValue const& per_thread) {
-    m_team_scratch_size[level]   = per_team.value;
-    m_thread_scratch_size[level] = per_thread.value;
-    return *this;
-  }
-
-  using member_type = Kokkos::Impl::SYCLTeamMember;
-
- protected:
-  template <class FunctorType>
-  int internal_team_size_max_for(const FunctorType& /*f*/) const {
-    // nested_reducer_memsize = (sizeof(double) * (m_team_size + 2)
-    // custom: m_team_scratch_size[0] + m_thread_scratch_size[0] * m_team_size
-    // total:
-    // 2*sizeof(double)+m_team_scratch_size[0]
-    // + m_team_size(sizeof(double)+m_thread_scratch_size[0])
-    const int max_threads_for_memory =
-        (space().impl_internal_space_instance()->m_maxShmemPerBlock -
-         2 * sizeof(double) - m_team_scratch_size[0]) /
-        (sizeof(double) + m_thread_scratch_size[0]);
-    return std::min({
-             int(m_space.impl_internal_space_instance()->m_maxWorkgroupSize),
-      // FIXME_SYCL Avoid requesting too many registers on NVIDIA GPUs.
-#if defined(KOKKOS_IMPL_ARCH_NVIDIA_GPU)
-                 256,
-#endif
-                 max_threads_for_memory
-           }) /
-           impl_vector_length();
-  }
-
-  template <class ValueType, class FunctorType>
-  int internal_team_size_max_reduce(const FunctorType& f) const {
-    using Analysis =
-        FunctorAnalysis<FunctorPatternInterface::REDUCE, TeamPolicyInternal,
-                        FunctorType, ValueType>;
-    using value_type      = typename Analysis::value_type;
-    const int value_count = Analysis::value_count(f);
-
-    // nested_reducer_memsize = (sizeof(double) * (m_team_size + 2)
-    // reducer_memsize = sizeof(value_type) * m_team_size * value_count
-    // custom: m_team_scratch_size[0] + m_thread_scratch_size[0] * m_team_size
-    // total:
-    // 2*sizeof(double)+m_team_scratch_size[0]
-    // + m_team_size(sizeof(double)+sizeof(value_type)*value_count
-    //               +m_thread_scratch_size[0])
-    const int max_threads_for_memory =
-        (space().impl_internal_space_instance()->m_maxShmemPerBlock -
-         2 * sizeof(double) - m_team_scratch_size[0]) /
-        (sizeof(double) + sizeof(value_type) * value_count +
-         m_thread_scratch_size[0]);
-    return std::min<int>({
-             int(m_space.impl_internal_space_instance()->m_maxWorkgroupSize),
-      // FIXME_SYCL Avoid requesting too many registers on NVIDIA GPUs.
-#if defined(KOKKOS_IMPL_ARCH_NVIDIA_GPU)
-                 256,
-#endif
-                 max_threads_for_memory
-           }) /
-           impl_vector_length();
-  }
-
-  template <class FunctorType>
-  int internal_team_size_recommended_for(const FunctorType& f) const {
-    // FIXME_SYCL improve
-    return 1 << Kokkos::Impl::int_log2(internal_team_size_max_for(f));
-  }
-
-  template <class ValueType, class FunctorType>
-  int internal_team_size_recommended_reduce(const FunctorType& f) const {
-    // FIXME_SYCL improve
-    return 1 << Kokkos::Impl::int_log2(
-               internal_team_size_max_reduce<ValueType>(f));
-  }
-};
-
-template <typename FunctorType, typename... Properties>
-class ParallelFor<FunctorType, Kokkos::TeamPolicy<Properties...>,
-                  Kokkos::Experimental::SYCL> {
- public:
-  using Policy = TeamPolicyInternal<Kokkos::Experimental::SYCL, Properties...>;
-  using functor_type = FunctorType;
-  using size_type    = ::Kokkos::Experimental::SYCL::size_type;
-
- private:
-  using member_type   = typename Policy::member_type;
-  using work_tag      = typename Policy::work_tag;
-  using launch_bounds = typename Policy::launch_bounds;
-
-  FunctorType const m_functor;
-  Policy const m_policy;
-  size_type const m_league_size;
-  int m_team_size;
-  size_type const m_vector_size;
-  int m_shmem_begin;
-  int m_shmem_size;
-  sycl::device_ptr<char> m_global_scratch_ptr;
-  size_t m_scratch_size[2];
-  // Only let one ParallelFor/Reduce modify the team scratch memory. The
-  // constructor acquires the mutex which is released in the destructor.
-  std::scoped_lock<std::mutex> m_scratch_lock;
-  int m_scratch_pool_id = -1;
-
-  template <typename FunctorWrapper>
-  sycl::event sycl_direct_launch(const Policy& policy,
-                                 const FunctorWrapper& functor_wrapper,
-                                 const sycl::event& memcpy_event) const {
-    // Convenience references
-    const Kokkos::Experimental::SYCL& space = policy.space();
-    sycl::queue& q                          = space.sycl_queue();
-
-    auto parallel_for_event = q.submit([&](sycl::handler& cgh) {
-      // FIXME_SYCL accessors seem to need a size greater than zero at least for
-      // host queues
-      sycl::local_accessor<char, 1> team_scratch_memory_L0(
-          sycl::range<1>(
-              std::max(m_scratch_size[0] + m_shmem_begin, size_t(1))),
-          cgh);
-
-      // Avoid capturing *this since it might not be trivially copyable
-      const auto shmem_begin       = m_shmem_begin;
-      const size_t scratch_size[2] = {m_scratch_size[0], m_scratch_size[1]};
-      sycl::device_ptr<char> const global_scratch_ptr = m_global_scratch_ptr;
-
-      auto lambda = [=](sycl::nd_item<2> item) {
-        const member_type team_member(
-            team_scratch_memory_L0.get_pointer(), shmem_begin, scratch_size[0],
-            global_scratch_ptr + item.get_group(1) * scratch_size[1],
-            scratch_size[1], item, item.get_group_linear_id(),
-            item.get_group_range(1));
-        if constexpr (std::is_void<work_tag>::value)
-          functor_wrapper.get_functor()(team_member);
-        else
-          functor_wrapper.get_functor()(work_tag(), team_member);
-      };
-
-      static sycl::kernel kernel = [&] {
-        sycl::kernel_id functor_kernel_id =
-            sycl::get_kernel_id<decltype(lambda)>();
-        auto kernel_bundle =
-            sycl::get_kernel_bundle<sycl::bundle_state::executable>(
-                q.get_context(), std::vector{functor_kernel_id});
-        return kernel_bundle.get_kernel(functor_kernel_id);
-      }();
-      auto max_sg_size =
-          kernel
-              .get_info<sycl::info::kernel_device_specific::max_sub_group_size>(
-                  q.get_device());
-      auto final_vector_size = std::min<int>(m_vector_size, max_sg_size);
-      // FIXME_SYCL For some reason, explicitly enforcing the kernel bundle to
-      // be used gives a runtime error.
-      // cgh.use_kernel_bundle(kernel_bundle);
-
-#ifndef KOKKOS_IMPL_SYCL_USE_IN_ORDER_QUEUES
-      cgh.depends_on(memcpy_event);
-#else
-      (void)memcpy_event;
-#endif
-      cgh.parallel_for(
-          sycl::nd_range<2>(
-              sycl::range<2>(m_team_size, m_league_size * final_vector_size),
-              sycl::range<2>(m_team_size, final_vector_size)),
-          lambda);
-    });
-#ifndef KOKKOS_IMPL_SYCL_USE_IN_ORDER_QUEUES
-    q.ext_oneapi_submit_barrier(std::vector<sycl::event>{parallel_for_event});
-#endif
-    return parallel_for_event;
-  }
-
- public:
-  inline void execute() const {
-    if (m_league_size == 0) return;
-
-    auto& space = *m_policy.space().impl_internal_space_instance();
-    Kokkos::Experimental::Impl::SYCLInternal::IndirectKernelMem&
-        indirectKernelMem = space.get_indirect_kernel_mem();
-
-    auto functor_wrapper = Experimental::Impl::make_sycl_function_wrapper(
-        m_functor, indirectKernelMem);
-
-    sycl::event event = sycl_direct_launch(m_policy, functor_wrapper,
-                                           functor_wrapper.get_copy_event());
-    functor_wrapper.register_event(event);
-    space.register_team_scratch_event(m_scratch_pool_id, event);
-  }
-
-  ParallelFor(FunctorType const& arg_functor, Policy const& arg_policy)
-      : m_functor(arg_functor),
-        m_policy(arg_policy),
-        m_league_size(arg_policy.league_size()),
-        m_team_size(arg_policy.team_size()),
-        m_vector_size(arg_policy.impl_vector_length()),
-        m_scratch_lock(arg_policy.space()
-                           .impl_internal_space_instance()
-                           ->m_team_scratch_mutex) {
-    // FIXME_SYCL optimize
-    if (m_team_size < 0)
-      m_team_size =
-          m_policy.team_size_recommended(arg_functor, ParallelForTag{});
-
-    m_shmem_begin = (sizeof(double) * (m_team_size + 2));
-    m_shmem_size =
-        (m_policy.scratch_size(0, m_team_size) +
-         FunctorTeamShmemSize<FunctorType>::value(m_functor, m_team_size));
-    m_scratch_size[0] = m_shmem_size;
-    m_scratch_size[1] = m_policy.scratch_size(1, m_team_size);
-
-    // Functor's reduce memory, team scan memory, and team shared memory depend
-    // upon team size.
-    auto& space       = *m_policy.space().impl_internal_space_instance();
-    m_scratch_pool_id = space.acquire_team_scratch_space();
-    m_global_scratch_ptr =
-        static_cast<sycl::device_ptr<char>>(space.resize_team_scratch_space(
-            m_scratch_pool_id,
-            static_cast<ptrdiff_t>(m_scratch_size[1]) * m_league_size));
-
-    if (static_cast<int>(space.m_maxShmemPerBlock) <
-        m_shmem_size - m_shmem_begin) {
-      std::stringstream out;
-      out << "Kokkos::Impl::ParallelFor<SYCL> insufficient shared memory! "
-             "Requested "
-          << m_shmem_size - m_shmem_begin << " bytes but maximum is "
-          << space.m_maxShmemPerBlock << '\n';
-      Kokkos::Impl::throw_runtime_exception(out.str());
-    }
-
-    const auto max_team_size =
-        m_policy.team_size_max(arg_functor, ParallelForTag{});
-    if (m_team_size > m_policy.team_size_max(arg_functor, ParallelForTag{}))
-      Kokkos::Impl::throw_runtime_exception(
-          "Kokkos::Impl::ParallelFor<SYCL> requested too large team size. The "
-          "maximal team_size is " +
-          std::to_string(max_team_size) + '!');
-  }
-};
-
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
 template <class CombinedFunctorReducerType, class... Properties>
-class ParallelReduce<CombinedFunctorReducerType,
-                     Kokkos::TeamPolicy<Properties...>,
-                     Kokkos::Experimental::SYCL> {
+class Kokkos::Impl::ParallelReduce<CombinedFunctorReducerType,
+                                   Kokkos::TeamPolicy<Properties...>,
+                                   Kokkos::Experimental::SYCL> {
  public:
   using Policy = TeamPolicyInternal<Kokkos::Experimental::SYCL, Properties...>;
   using FunctorType = typename CombinedFunctorReducerType::functor_type;
@@ -951,7 +454,5 @@ class ParallelReduce<CombinedFunctorReducerType,
     initialize();
   }
 };
-}  // namespace Impl
-}  // namespace Kokkos
 
 #endif
